hw0_release/requirements.txt
bleach==2.0.0
cycler==0.10.0
decorator==4.0.11
entrypoints==0.2.3
html5lib==0.999999999
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.0
jupyter-console==5.1.0
jupyter-core==4.4.0
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
networkx==1.11
notebook==5.4.1
numpy==1.16.2
olefile==0.44
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==2.0.4
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
pyzmq==18.0.1
qtconsole==4.3.0
scikit-image==0.14.0
scipy==1.2.1
simplegeneric==0.8.1
six==1.10.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
